refactor: trace.rs -> segment.rs
