+++ libdd-trace-utils/src/change_buffer/segment.rs
